# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Quoc Sach Driver booking site
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type City = (typeof City)[keyof typeof City];

export const City = {
  "da-nang": "da-nang",
  hue: "hue",
  "hoi-an": "hoi-an",
} as const;

export type LocationCategory =
  (typeof LocationCategory)[keyof typeof LocationCategory];

export const LocationCategory = {
  beach: "beach",
  heritage: "heritage",
  temple: "temple",
  mountain: "mountain",
  food: "food",
  village: "village",
  landmark: "landmark",
  museum: "museum",
} as const;

export interface Location {
  id: string;
  slug: string;
  city: City;
  category: LocationCategory;
  nameEn: string;
  nameVi: string;
  nameKo: string;
  descriptionEn: string;
  descriptionVi: string;
  descriptionKo: string;
  imageUrl: string;
  lat: number;
  lng: number;
  featured: boolean;
  /** Suggested visit duration in minutes */
  durationMinutes: number;
}

export type RateCardCitySurchargesItem = {
  city: City;
  surchargeUsd: number;
};

export interface RateCard {
  /** Base price per day for the driver service in USD */
  baseDailyPriceUsd: number;
  currency: string;
  citySurcharges: RateCardCitySurchargesItem[];
  /** Fee per person beyond the first 4 */
  extraPersonFeeUsd: number;
  /** People threshold above which the per-person fee applies */
  extraPersonThreshold: number;
}

export interface PriceEstimateRequest {
  startDate: string;
  endDate: string;
  /** @minimum 1 */
  peopleCount: number;
  cities: City[];
}

export interface PriceLineItem {
  label: string;
  amountUsd: number;
}

export interface PriceEstimate {
  days: number;
  peopleCount: number;
  baseTotalUsd: number;
  cityTotalUsd: number;
  peopleTotalUsd: number;
  totalUsd: number;
  currency: string;
  lineItems: PriceLineItem[];
}

export interface ItineraryStop {
  locationId: string;
  order: number;
}

export interface ItineraryDay {
  day: number;
  stops: ItineraryStop[];
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  startDate: string;
  endDate: string;
  peopleCount: number;
  cities: City[];
  itinerary: ItineraryDay[];
  message?: string;
  totalUsd: number;
  status: BookingStatus;
  language?: string;
  createdAt: string;
}

export interface BookingInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  /** @minLength 1 */
  customerPhone: string;
  startDate: string;
  endDate: string;
  /** @minimum 1 */
  peopleCount: number;
  cities: City[];
  itinerary: ItineraryDay[];
  message?: string;
  language?: string;
}

export interface BookingStatusPatch {
  status: BookingStatus;
}

export interface Review {
  id: string;
  author: string;
  country: string;
  rating: number;
  title: string;
  body: string;
  avatarUrl?: string;
  createdAt: string;
}

export interface StatusCount {
  status: BookingStatus;
  count: number;
}

export interface AdminSummary {
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  completedBookings: number;
  upcomingTrips: number;
  totalRevenueUsd: number;
  recentBookings: Booking[];
  statusBreakdown: StatusCount[];
}

export type ListLocationsParams = {
  city?: ListLocationsCity;
};

export type ListLocationsCity =
  (typeof ListLocationsCity)[keyof typeof ListLocationsCity];

export const ListLocationsCity = {
  "da-nang": "da-nang",
  hue: "hue",
  "hoi-an": "hoi-an",
} as const;

export type ListBookingsParams = {
  status?: ListBookingsStatus;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;