# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Quoc Sach Driver booking site
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { City } from "./city";
import type { ItineraryDay } from "./itineraryDay";

export interface Booking {
  id: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  startDate: Date;
  endDate: Date;
  peopleCount: number;
  cities: City[];
  itinerary: ItineraryDay[];
  message?: string;
  totalUsd: number;
  status: BookingStatus;
  language?: string;
  createdAt: Date;
}