# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Quoc Sach Driver booking site
 * OpenAPI spec version: 0.1.0
 */
import type { City } from "./city";
import type { ItineraryDay } from "./itineraryDay";

export interface BookingInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  /** @minLength 1 */
  customerPhone: string;
  startDate: Date;
  endDate: Date;
  /** @minimum 1 */
  peopleCount: number;
  cities: City[];
  itinerary: ItineraryDay[];
  message?: string;
  language?: string;
}