# lib/api-zod/src/generated/types/itineraryDay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Quoc Sach Driver booking site
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryStop } from "./itineraryStop";

export interface ItineraryDay {
  day: number;
  stops: ItineraryStop[];
}