# lib/api-zod/src/generated/types/locationCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Quoc Sach Driver booking site
 * OpenAPI spec version: 0.1.0
 */

export type LocationCategory =
  (typeof LocationCategory)[keyof typeof LocationCategory];

export const LocationCategory = {
  beach: "beach",
  heritage: "heritage",
  temple: "temple",
  mountain: "mountain",
  food: "food",
  village: "village",
  landmark: "landmark",
  museum: "museum",
} as const;