# lib/api-zod/src/generated/types/priceEstimate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Quoc Sach Driver booking site
 * OpenAPI spec version: 0.1.0
 */
import type { PriceLineItem } from "./priceLineItem";

export interface PriceEstimate {
  days: number;
  peopleCount: number;
  baseTotalUsd: number;
  cityTotalUsd: number;
  peopleTotalUsd: number;
  totalUsd: number;
  currency: string;
  lineItems: PriceLineItem[];
}