# lib/api-zod/src/generated/types/priceLineItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Quoc Sach Driver booking site
 * OpenAPI spec version: 0.1.0
 */

export interface PriceLineItem {
  label: string;
  amountUsd: number;
}